pyproject.toml
[project]
name = "floki-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.4",
    "chromadb>=1.0.20",
    "fastapi>=0.116.1",
    "openai-agents>=0.2.9",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
